# project/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BudgetBuddy API Documentation",
        "description": "API pour la gestion des dépenses personnelles et des tags",
        "contact": {
            "name": "Support API BudgetBuddy",
            "email": "contact@budgetbuddy.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "BudgetBuddy API Server"
        }
    ],
    "paths": {
...
        "/api/expenses": {
            "get": {
                "tags": [
                    "Expenses"
                ],
                "summary": "Liste toutes les dépenses de l'utilisateur authentifié",
                "operationId": "8a0ec0f16f4930b3b9220615e6ef4329",
                "responses": {
                    "200": {
                        "description": "Liste des dépenses",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "Courses supermarché"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Achats hebdomadaires"
                                            },
                                            "amount": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 85.5
                                            },
                                            "expense_date": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2025-03-12"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "tags": {
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Alimentation"
                                                        },
                                                        "color": {
                                                            "type": "string",
                                                            "example": "#FF5733"
                                                        },
                                                        "created_at": {
                                                            "type": "string",
                                                            "format": "date-time"
                                                        },
                                                        "updated_at": {
                                                            "type": "string",
                                                            "format": "date-time"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Expenses"
                ],
                "summary": "Crée une nouvelle dépense",
                "operationId": "91b5f56eb459bb69ea5cb599da73e40f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "amount",
                                    "expense_date"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Courses supermarché"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Achats hebdomadaires"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 85.5
                                    },
                                    "expense_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-03-12"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Dépense créée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Courses supermarché"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Achats hebdomadaires"
                                        },
                                        "amount": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 85.5
                                        },
                                        "expense_date": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2025-03-12"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Données invalides",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/expenses/{id}/tags": {
            "post": {
                "tags": [
                    "Expenses"
                ],
                "summary": "Associe des tags à une dépense",
                "operationId": "c20fbb506af316940d495a948af1d9c5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la dépense",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "tags"
                                ],
                                "properties": {
                                    "tags": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            2
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tags associés avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Courses supermarché"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Achats hebdomadaires"
                                        },
                                        "amount": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 85.5
                                        },
                                        "expense_date": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2025-03-12"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "tags": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Alimentation"
                                                    },
                                                    "color": {
                                                        "type": "string",
                                                        "example": "#FF5733"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Données invalides",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Dépense non trouvée"
                    },
                    "403": {
                        "description": "Accès interdit"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}